refactor: HttpRequest 외부 객체로 분리
