+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -9,7 +9,6 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.Socket;
 import java.nio.file.Files;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import org.apache.coyote.Processor;
@@ -31,58 +30,6 @@ public class Http11Processor implements Runnable, Processor {
     public void run() {
         log.info("connect host: {}, port: {}", connection.getInetAddress(), connection.getPort());
         process(connection);
-    }
-
-    static class HttpRequest {
-        private final String method;
-        private final String uri;
-
-        public HttpRequest(String method, String uri) {
-            this.method = method;
-            this.uri = uri;
-        }
-
-        private Optional<String> getQueryString() {
-            int index = uri.indexOf("?");
-            if (index == -1) {
-                return Optional.empty();
-            }
-
-            return Optional.of(uri.substring(index + 1));
-        }
-
-        public Map<String, String> getQueryStrings() {
-            Map<String, String> queryStrings = new HashMap<>();
-
-            Optional<String> queryString = getQueryString();
-            if (queryString.isEmpty()) {
-                return new HashMap<>();
-            }
-
-            for (String keyValue : queryString.get().split("&")) {
-                String[] keyValues = keyValue.split("=");
-                queryStrings.put(keyValues[0], keyValues[1]);
-            }
-
-            return queryStrings;
-        }
-
-        public String getPath() {
-            int index = uri.indexOf("?");
-            if (index == -1) {
-                return uri;
-            }
-
-            return uri.substring(0, index);
-        }
-
-        public String getMethod() {
-            return method;
-        }
-
-        public String getUri() {
-            return uri;
-        }
     }
 
     @Override
